+++ docs/about/changelog.md
@@ -13,6 +13,7 @@ toc_depth: 2
 
 - version init argument typing #401
 - include `lib_name` and `lib_version` in `FakeBaseConnectionMixin`
+- when using `FakeValkey`, raise `valkey.ResponseError` instead of `redis.ResponseError` #402
 
 ## v2.30.3 - 2025-08-01